# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexgohost Billing & Client Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
  /admin/clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /admin/clients/{id}/suspend:
    post:
      operationId: suspendClient
      tags: [clients]
      summary: Suspend client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /admin/clients/{id}/activate:
    post:
      operationId: activateClient
      tags: [clients]
      summary: Activate client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /admin/clients/{id}/reset-password:
    post:
      operationId: resetClientPassword
      tags: [clients]
      summary: Reset client password
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /account:
    get:
      operationId: getAccount
      tags: [auth]
      summary: Get account settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateAccount
      tags: [auth]
      summary: Update account settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        company:
          type: string
        phone:
          type: string
      required:
        - firstName
        - lastName
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, client]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - status
        - createdAt
    UpdateAccountRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phone:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string
    UpdateClientRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        status:
          type: string
    ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
...
    Client:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
        servicesCount:
          type: number
        domainsCount:
          type: number
      required:
        - id
        - firstName
        - lastName
        - email
        - status
        - createdAt
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            hosting:
              type: array
              items:
                $ref: "#/components/schemas/HostingService"
            domains:
              type: array
              items:
                $ref: "#/components/schemas/Domain"
            invoices:
              type: array
              items:
                $ref: "#/components/schemas/Invoice"
            tickets:
              type: array
              items:
                $ref: "#/components/schemas/Ticket"
    ClientListResponse:
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - clients
        - total
        - page
        - limit
...
    HostingService:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        planId:
          type: string
        planName:
          type: string
        domain:
          type: string
        username:
          type: string
        serverIp:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated, pending]
        startDate:
          type: string
        expiryDate:
          type: string
        diskUsed:
          type: string
        bandwidthUsed:
          type: string
        cpanelUrl:
          type: string
      required:
        - id
        - clientId
        - planId
        - planName
        - status
    Domain:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        name:
          type: string
        tld:
          type: string
        registrationDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired, pending, transferred]
        autoRenew:
          type: boolean
        nameservers:
          type: array
          items:
            type: string
      required:
        - id
        - clientId
        - name
        - tld
        - status
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [unpaid, paid, cancelled, overdue]
        dueDate:
          type: string
        paidDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - clientId
        - amount
        - total
        - status
        - dueDate
        - createdAt
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - total
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [open, closed, pending, answered]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        department:
          type: string
        lastReply:
          type: string
        messagesCount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - ticketNumber
        - clientId
        - subject
        - status
        - priority
        - createdAt
...