# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexgohost Billing & Client Management API
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from "./domainStatus";

export interface Domain {
  id: string;
  clientId: string;
  clientName?: string;
  name: string;
  tld: string;
  registrationDate?: string;
  expiryDate?: string;
  status: DomainStatus;
  autoRenew?: boolean;
  nameservers?: string[];
}